requirements.txt
# requirements.txt
pyspark==3.5.1
pandas==2.2.2
minio==7.1.17
boto3==1.34.136
pyarrow==16.1.0
pendulum==3.0.0
psycopg2-binary==2.9.9
tenacity==8.2.3
pytest==7.4.4
fsspec==2023.10.0